test(runtime): separate extension registry coverage
